add: opcodes faltantes e realização de correções em timers e ciclos da cpu
Falta ainda:
  - Corrigir problema de excesso de velocidade do jogo
  - Corrigir problema no display que fica chuviscando estranhamente

+++ src/include/timer.h
@@ -7,5 +7,5 @@ extern const double CPU_INTERVAL;
 extern double timer_accumulator;
 extern double cpu_accumulator;
 
-extern Uint32 last_ticks;
-extern Uint32 cpu_last_ticks;
+extern Uint64 last_ticks;
+extern Uint64 cpu_last_ticks;
