+++ src/timer.c
@@ -6,5 +6,5 @@ const double CPU_INTERVAL = 1.0 / 700.0;
 double timer_accumulator = 0.0;
 double cpu_accumulator = 0.0;
 
-Uint32 last_ticks = 0;
-Uint32 cpu_last_ticks = 0;
+Uint64 last_ticks = 0;
+Uint64 cpu_last_ticks = 0;
